+++ testFiles/requirejs-config.js
@@ -1,0 +1,9 @@
+var config = {
+  config: {
+    mixins: {
+      mixinTarget: {
+        mixinNavigation: true
+      }
+    }
+  }
+}